互換性向上

+++ src/random.h
@@ -2,7 +2,7 @@
 #define __RANDOM_H__
 
 extern void srrandom(int seed);
-extern long rrandom(void);
+extern unsigned short rrandom(void);
 extern int get_rand(int low, int high);
 extern int rand_percent(int percentage);
 extern int coin_toss(void);
