+++ src/room.c
@@ -30,7 +30,7 @@ void light_up_room(int rn)
         (rooms[rn].is_room & R_MAZE)) return;
     for (i = rooms[rn].top_row; i <= rooms[rn].bottom_row; i++) {
         for (j = rooms[rn].left_col; j <= rooms[rn].right_col; j++) {
-            colorize_dungeon((short)i, (short)j);
+            colorize_dungeon(i, j);
         }
     }
     attrset(A_NORMAL);
@@ -47,7 +47,7 @@ void light_passage(int row, int col)
     for (i = ((row > MIN_ROW) ? -1 : 0); i <= i_end; i++) {
         for (j = ((col > 0) ? -1 : 0); j <= j_end; j++) {
             if (can_move(row, col, row + i, col + j)) {
-                colorize_dungeon((short)(row + i), (short)(col + j));
+                colorize_dungeon(row + i, col + j);
             }
         }
     }
@@ -63,7 +63,7 @@ void darken_room(short rn)
     for (i = rooms[rn].top_row + 1; i < rooms[rn].bottom_row; i++) {
         for (j = rooms[rn].left_col + 1; j < rooms[rn].right_col; j++) {
             if (blind || (DUNGEON(i, j) != TILE_STAIRS &&
-                !object_at(&level_objects, (short)i, (short)j))) {
+                !object_at(&level_objects, i, j))) {
                 DUNGEON_ATTR(i, j) = ATTR_HIDDEN;
             }
         }
@@ -77,8 +77,8 @@ gr_row_col(short *row, short *col, unsigned short mask)
     u8 tile;
 
     do {
-        *row = (short)get_rand(MIN_ROW, MAX_ROW);
-        *col = (short)get_rand(0, ROGUE_COLUMNS - 1);
+        *row = get_rand(MIN_ROW, MAX_ROW);
+        *col = get_rand(0, ROGUE_COLUMNS - 1);
         tile = DUNGEON(*row, *col);
     } while (!((tile == TILE_FLOOR && (mask & FLOOR)) ||
                (tile == TILE_TUNNEL && (mask & TUNNEL)) ||
@@ -103,15 +103,15 @@ int party_objects(int rn)
     short row, col;
     object *obj;
 
-    n = (u8)get_rand(5, 10);
+    n = get_rand(5, 10);
     i = 0;
     do {    // z88dk メモリ削減対策(while->doに変更)
         obj = gr_object();
         if (!obj) continue;
         j = 0;
         do {    // z88dk メモリ削減対策(while->doに変更)
-            row = (short)get_rand(rooms[rn].top_row + 1, rooms[rn].bottom_row - 1);
-            col = (short)get_rand(rooms[rn].left_col + 1, rooms[rn].right_col - 1);
+            row = get_rand(rooms[rn].top_row + 1, rooms[rn].bottom_row - 1);
+            col = get_rand(rooms[rn].left_col + 1, rooms[rn].right_col - 1);
             if ((DUNGEON(row,col) == TILE_FLOOR || DUNGEON(row,col) == TILE_TUNNEL) &&
                 !object_at(&level_objects, row, col)) {
                 place_at(obj, row, col);
